Remove all population based model code
