+++ src/wolves_in_silico/abm/player.py
@@ -1,4 +1,9 @@
-from wolves_in_silico.base.game import Role
+from enum import Enum
+
+
+class Role(Enum):
+    WOLF = 1
+    CIV = 2
 
 
 class Player:
